Add files via upload
For this game, you want to download the files as a zip, and then extract them. Put the photos into a folder, put the music into a folder, and then make a new folder in which will hold all of your files. Finally, put the index.html into your browser, and click play!